Deliverable2_ver0.5
Everything related to login and player records is fully functional.

Written some partial code for printing questions.

Note on the keywords in the comments: (Ctrl + F to look for these keywords reccommended)
TODO: highlights the sections of the code that require completion
NEWCODE: exclusive to Application.java. Easier to navigate to the lines that are different from deliverable1.
IDEA: general ideas